Moved CfgSounds to its own file
