+++ addons/main/CfgSounds.hpp
@@ -1,0 +1,13 @@
+class CfgSounds {
+	sounds[] = {};
+	class ade_sound_zischen {
+		name = "Zischen";
+		sound[] = {QPATHTOF(data\sounds\zischen.ogg), 1, 1};
+		titles[] = {};
+	};
+	class ade_sound_beep {
+		name = "Beep";
+		sound[] = {QPATHTOF(data\sounds\beep.ogg), 1, 1};
+		titles[] = {};
+	};
+};
