+++ addons/main/config.cpp
@@ -21,17 +21,4 @@ class Extended_PostInit_EventHandlers {
 #include "CfgGlasses.hpp"
 #include "CfgVehicles.hpp"
 #include "CfgWeapons.hpp"
-
-class CfgSounds {
-	sounds[] = {};
-	class ade_sound_zischen {
-		name = "Zischen";
-		sound[] = {QPATHTOF(data\sounds\zischen.ogg), 1, 1};
-		titles[] = {};
-	};
-	class ade_sound_beep {
-		name = "Beep";
-		sound[] = {QPATHTOF(data\sounds\beep.ogg), 1, 1};
-		titles[] = {};
-	};
-};
+#include "CfgSounds.hpp"
